+++ contracts/IVoucherHub.sol
@@ -16,6 +16,7 @@ interface IVoucherHub {
         uint256 value
     );
     event VoucherDebited(address indexed voucher, uint256 sponsoredAmount);
+    event VoucherRefunded(address indexed voucher, uint256 amount);
     event VoucherTypeCreated(uint256 indexed id, string description, uint256 duration);
     event VoucherTypeDescriptionUpdated(uint256 indexed id, string description);
     event VoucherTypeDurationUpdated(uint256 indexed id, uint256 duration);
@@ -39,8 +40,10 @@ interface IVoucherHub {
         address dataset,
         uint256 datasetPrice,
         address workerpool,
-        uint256 workerpoolPrice
+        uint256 workerpoolPrice,
+        uint256 volume
     ) external returns (uint256 sponsoredAmount);
+    function refundVoucher(uint256 amount) external;
 
     function getIexecPoco() external view returns (address);
     function getVoucherBeacon() external view returns (address);